Added wasm compatible logger implementation
