+++ lib/src/misc/logger/index.dart
@@ -1,0 +1,2 @@
+export 'logger.dart';
+export 'io.dart' if (dart.library.js_interop) 'web.dart';
